+++ Gallery/trunk/src/org/shinshi/gallery/dao/RemarkDao.java
@@ -1,5 +1,4 @@
 package org.shinshi.gallery.dao;
-import VO.Postmodel;
 import VO.Remarkids;
 
 import java.util.List;
@@ -9,5 +8,5 @@ public interface RemarkDao extends IBaseDao<Remarkids,Integer>{
 	public List<Integer> getRemarkbyPostmodel(Integer PostId);
 	public List<Remarkids> getRemarkbyUsermodel(Integer UserId);
 	public List<Integer> getRemarkbyUsermodelandPostmodel(Integer UserId,Integer PostId);
-	public Integer getAverageScore(Integer PostId);
+	public Float getAverageScore(Integer PostId);
 }
